requirements.txt
`requirements.txt` (环境清单)

```text
# --- Hanasuki Beta 1.1 Core Dependencies ---
PyYAML>=6.0.1
PyQt6>=6.6.1
numpy>=1.26.0
networkx>=3.2.1

# --- LLM Backend (HERO-A+ Optimization) ---
# 核心推理引擎
llama-cpp-python>=0.2.26

# --- Memory Engine (Semantic + Rational) ---
# 双轨存储
lancedb>=0.4.0
pyarrow>=14.0.0
sentence-transformers>=2.2.2
torch>=2.2.0

# --- Web & Tools (Purify Search) ---
# 学术净化浏览器
playwright>=1.41.0
beautifulsoup4>=4.12.2

# --- UI & Formatting ---
# Markdown 渲染
markdown2>=2.4.13

# --- System Utils ---
# 主权校验
binascii